// packages/eorm-wizard/src/generated/EormSetup.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EPackage, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { EntityConfig } from './EntityConfig';

/**
 * EormSetup
 * @generated
 */
export interface EormSetup extends EObject {
  mappingName: string;
  schema?: string;
  modelPackage: EPackage;
  delimitedIdentifiers?: boolean;
  entities: EntityConfig[];

}