// packages/eorm-wizard/src/generated/InheritanceStrategy.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

/**
 * InheritanceStrategy
 * @generated
 */
export const InheritanceStrategy = {
  NONE: 'NONE',
  SINGLE_TABLE: 'SINGLE_TABLE',
  JOINED: 'JOINED',
  TABLE_PER_CLASS: 'TABLE_PER_CLASS'
} as const;

export type InheritanceStrategy = typeof InheritanceStrategy[keyof typeof InheritanceStrategy];

/**
 * Get enum value by name
 */
export function getInheritanceStrategy(name: string): InheritanceStrategy | undefined {
  return (InheritanceStrategy as Record<string, string>)[name] as InheritanceStrategy | undefined;
}

/**
 * Get enum value by literal
 */
export function getInheritanceStrategyByLiteral(literal: string): InheritanceStrategy | undefined {
  if (literal === 'NONE') return InheritanceStrategy.NONE;
  if (literal === 'SINGLE_TABLE') return InheritanceStrategy.SINGLE_TABLE;
  if (literal === 'JOINED') return InheritanceStrategy.JOINED;
  if (literal === 'TABLE_PER_CLASS') return InheritanceStrategy.TABLE_PER_CLASS;
  return undefined;
}